Add files via upload
Online SB detection is a prevalent crime and very difficult to detect because of its very similar 
behavior during real-time bidding. Due to SB, the other bidder gets a lot of money loss, and 
the seller receives the extra money. As defined earlier, the previous researcher has a lot of 
work, but there is always a gap in research. So, this article developed a decision base fusion 
model used to find the SB in the real-time auction in which SVM and ANN are used for 
prediction, and Fuzzy is used to decide whether the SB is committed or the bid is normal. 
When the bid is made the bidding, the behavior is judge by both SVM and ANN at the same 
time. Based on their prediction, fuzzy decide the bid is normal or SB. 
The proposed model predicts the best results compared to the previous research models, which 
help detect real-time auction fraud, which helps to block the user and discard fake bids. 
Therefore, this research will be helpful to both bidders and the e-Auction companies that face 
the yearly millions of dollars loss and fraud reports.